backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
openai==1.14.3
python-multipart==0.0.9
python-dotenv==1.0.1
pymongo==4.6.1
bcrypt==4.1.2
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0